sigma-runtime/SR-EI-037/code/requirements.txt
# SIGMA Runtime v0.3.7 - Dependencies
# Python 3.10+

openai>=1.0.0
sentence-transformers>=2.2.0
scipy>=1.10.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
torch>=2.0.0
